+++ Source/ACE.Database/Models/Shard/Character.cs
@@ -12,6 +12,11 @@ namespace ACE.Database.Models.Shard
         public bool IsDeleted { get; set; }
         public ulong DeleteTime { get; set; }
         public double LastLoginTimestamp { get; set; }
+        public int TotalLogins { get; set; }
+        public int CharacterOptions1 { get; set; }
+        public int CharacterOptions2 { get; set; }
+        public uint HairTexture { get; set; }
+        public uint DefaultHairTexture { get; set; }
 
         public Biota Biota { get; set; }
     }