+++ Source/ACE.Database/Models/Shard/CharacterPropertiesQuestRegistry.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace ACE.Database.Models.Shard
 {
-    public partial class BiotaPropertiesQuestRegistry
+    public partial class CharacterPropertiesQuestRegistry
     {
         public uint Id { get; set; }
         public uint ObjectId { get; set; }