+++ Source/ACE.Database/Models/Shard/CharacterPropertiesShortcutBar.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 
 namespace ACE.Database.Models.Shard
 {
-    public partial class BiotaPropertiesTitleBook
+    public partial class CharacterPropertiesShortcutBar
     {
         public uint Id { get; set; }
         public uint ObjectId { get; set; }
-        public uint TitleId { get; set; }
+        public uint ShortcutBarIndex { get; set; }
+        public uint ShortcutObjectId { get; set; }
 
         public Biota Object { get; set; }
     }