+++ Source/ACE.Database/Models/Shard/CharacterPropertiesSpellBar.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 
 namespace ACE.Database.Models.Shard
 {
-    public partial class BiotaPropertiesFillCompBook
+    public partial class CharacterPropertiesSpellBar
     {
         public uint Id { get; set; }
         public uint ObjectId { get; set; }
-        public int SpellComponentId { get; set; }
-        public int QuantityToRebuy { get; set; }
+        public uint SpellBarNumber { get; set; }
+        public uint SpellBarIndex { get; set; }
+        public uint SpellId { get; set; }
 
         public Biota Object { get; set; }
     }