+++ Source/ACE.Database/Models/Shard/CharacterPropertiesTitleBook.cs
@@ -3,14 +3,12 @@ using System.Collections.Generic;
 
 namespace ACE.Database.Models.Shard
 {
-    public partial class BiotaPropertiesShortcutBar
+    public partial class CharacterPropertiesTitleBook
     {
         public uint Id { get; set; }
         public uint ObjectId { get; set; }
-        public uint ShortcutBarIndex { get; set; }
-        public uint ShortcutObjectId { get; set; }
+        public uint TitleId { get; set; }
 
         public Biota Object { get; set; }
-        public Biota ShortcutObject { get; set; }
     }
 }